[fontir] Add ir::path_builder module
And move all the GlyphPathBuilder stuff there.

No functional change. I'd like to hang the 'erase open corners' code off
of this, and the ir module is already huge.
